Extract content writing logic into ContentWriteExecutor
- This is the first step to add a queue between the controller and
the actual writing of content to the DB
